requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
dj-database-url==1.0.0
Django==4.1.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0
requests==2.28.1
Pillow==9.3
Django==3.2.9
matplotlib==3.4.3
opencv-python==4.5.4.58
numpy==1.21.3
imutils==0.5.4
easyocr==1.4
Pillow==8.4.0